add code clone detection

+++ se_tasks/clone_detection/scripts/main.py
@@ -77,7 +77,7 @@ def test_model(val_loader, model, device):
         node_1, graph_1, node_2, graph_2, label = data
         label = torch.tensor(label, dtype=torch.float, device=device)
         output = model(node_1, graph_1, node_2, graph_2, device)
-        acc += ((output > 0) == label).sum()
+        acc += ((output > 0) == label).sum().float()
     acc = acc / len(val_loader.dataset)
     return acc
